% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue.R
\docType{package}
\name{bluepally}
\alias{bluepally}
\alias{blueberry_colors}
\alias{scale_color_blueberry}
\alias{scale_fill_blueberry}
\alias{blueberry_cols}
\title{These functions provide colour palettes based on fresh and delicious blueberries.}
\format{
An object of class \code{character} of length 8.
}
\usage{
scale_color_blueberry(palette = "main", reverse = FALSE, ...)
scale_fill_blueberry(palette = "main", reverse = FALSE, ...)
blueberry_cols()
}
\arguments{
\item{ncol}{the number of colors you want}
}
\examples{
library(ggplot2)

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
geom_point(size = 4, alpha = .6) +
scale_color_blueberry(discrete = FALSE, palette = "blue")

ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
geom_bar() +
theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
scale_fill_blueberry( guide = "none")

blueberry_pal("blue")(10)
}
\keyword{datasets}
